Infrastructure for multiple whole-cell skipping decision functions
While the is plenty of options to clear the output / metadata of a cell,
there is only one defined way of skipping an entire cell: If the cell is
empty. This commit adds the infrastructure for including more conditions
that determine whether the cell (not only output, but source as well)
will be skipped.

The Signature of _cell() changes from a function as a second parameter to
a list of functions