Add an --drop-tagged-cells="some tags" Option (#161)
* Infrastructure for multiple whole-cell skipping decision functions

While the is plenty of options to clear the output / metadata of a cell,
there is only one defined way of skipping an entire cell: If the cell is
empty. This commit adds the infrastructure for including more conditions
that determine whether the cell (not only output, but source as well)
will be skipped.

The Signature of _cell() changes from a function as a second parameter to
a list of functions

* Add an apparatus for stripping cells from a Notebook

Introduce --drop-tagged-cells="tag1 tag2 ..." option. Given a notebook,
each cell, that has one of these tags will be removed entirely from the
notebook. This is useful for students exercises. Flagging a cell as e.g.
solution can help you removing them in an automated fashion from e.g. a
makefile.

Rename option from stripped to dropped