Mark git filter required on installation
By [default](https://git-scm.com/docs/gitattributes#_filter), "A missing filter driver definition in the config, or a filter driver that exits with a non-zero status, is not an error but makes the filter a no-op passthru." i.e. if the `nbstripout` is misconfigured or fails, the filter is ignored.

Declare the filter to be `required` when installing, such that it *must* succeed.

Closes #191